+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Support for `.isPortrait`, `.setPortrait` and `setLandscape` on Device.
 
+### Removed
+
+- Support for `.setOrientation` on Device.
+
 ## [v1.0.0-alpha.15] - 2019-12-23
 
 ### Added